+++ src/Goliath.Utilities/Authorization/PermissionList.cs
@@ -4,7 +4,7 @@ namespace Goliath.Authorization
     /// 
     /// </summary>
     [System.Serializable]
-    public class PermissionList : System.Collections.ObjectModel.KeyedCollection<int, IPermissionItem>
+    public class PermissionList : System.Collections.ObjectModel.KeyedCollection<long, IPermissionItem>
     {
         /// <summary>
         /// When implemented in a derived class, extracts the key from the specified element.
@@ -13,7 +13,7 @@ namespace Goliath.Authorization
         /// <returns>
         /// The key for the specified element.
         /// </returns>
-        protected override int GetKeyForItem(IPermissionItem item)
+        protected override long GetKeyForItem(IPermissionItem item)
         {
             return item.RoleNumber;
         }
